+++ packages/image-block/src/types.ts
@@ -102,10 +102,3 @@ export const mapCaptionPositionClasses: Record<CaptionPosition, string> = {
     [CaptionPosition.Right]: 'tw-flex-col md:!tw-flex-row',
     [CaptionPosition.Left]: 'tw-flex-col md:!tw-flex-row-reverse',
 };
-
-export enum ImageFormat {
-    WEBP = 'webp',
-    JPG = 'jpg',
-    PNG = 'png',
-    GIF = 'gif',
-}
