+++ packages/shared/src/components/ResponsiveImage/index.tsx
@@ -1,5 +1,3 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
-export * from './CssValueDisplay';
-export * from './EditAltTextFlyout';
 export * from './ResponsiveImage';
