+++ packages/shared/src/index.ts
@@ -1,3 +1,5 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
 export * from './components';
+export * from './hooks';
+export * from './types';
