+++ packages/shared/src/types/index.ts
@@ -1,0 +1,8 @@
+/* (c) Copyright Frontify Ltd., all rights reserved. */
+
+export enum ImageFormat {
+    WEBP = 'webp',
+    JPG = 'jpg',
+    PNG = 'png',
+    GIF = 'gif',
+}
